# backend/StigviddAPI/Controllers/TrailObstaclesController.cs
// SPDX-FileCopyrightText: 2025-2026 The Stigvidd Authors
// SPDX-License-Identifier: AGPL-3.0-or-later

using Core.Interfaces.Services;
using Infrastructure.Enums;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using WebDataContracts.RequestModels.TrailObstacle;
using WebDataContracts.ResponseModels.TrailObstacle;


namespace StigviddAPI.Controllers;

[ApiController]
[Route("api/v1/[controller]")]
public class TrailObstaclesController : StigViddController
{
    private readonly ITrailObstaclesService _obstaclesService;
    private readonly IUserService _userService;
    private readonly ILogger<TrailObstaclesController> _logger;

    public TrailObstaclesController(ITrailObstaclesService obstaclesService, IUserService userService, ILogger<TrailObstaclesController> logger)
    {
        _obstaclesService = obstaclesService;
        _userService = userService;
        _logger = logger;
    }

    [AllowAnonymous]
    [HttpGet]
    [Route("trail/{trailIdentifier}")]
    public async Task<ActionResult<IReadOnlyCollection<TrailObstacleResponse>>> GetTrailObstacles(
         [FromRoute] string trailIdentifier,
        CancellationToken ctoken)
    {
        var result = await _obstaclesService.GetTrailObstaclesByTrailIdentifierAsync(trailIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation(
                "GetTrailObstacles: Failed to fetch obstacles for trail with identifier: {identifier}.", trailIdentifier);

            return ToActionResult(result.Message);
        }

        return Ok(result.Value);
    }

    [AllowAnonymous]
    [HttpGet]
    [Route("issue-types")]
    public ActionResult<IReadOnlyCollection<string>> GetTrailObstacleIssueTypes()
    {
        return Ok(Enum.GetNames<TrailIssueType>());
    }

    [Authorize]
    [HttpPost]
    [Route("")]
    public async Task<ActionResult> CreateTrailObstacle(
       TrailObstacleRequest obstacleRequest,
        CancellationToken ctoken)
    {
        var user = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (user is null)
        {
            return Unauthorized("User not found");
        }

        var result = await _obstaclesService.AddTrailObstacle(
            user.Identifier,
            obstacleRequest.TrailIdentifier,
            obstacleRequest.Description,
            obstacleRequest.IssueType,
            obstacleRequest.IncidentLongitude,
            obstacleRequest.IncidentLatitude,
            ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("AddTrailObstacle: Failed to save trail obstacle for user {user}", user.Identifier);
            return ToActionResult(result.Message);
        }

        return Created();
    }

    [Authorize]
    [HttpPost]
    [Route("solve/{trailObstacleIdentifier}")]
    public async Task<ActionResult> AddSolvedVote([FromRoute] string trailObstacleIdentifier, CancellationToken ctoken)
    {
        var user = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (user is null)
        {
            return Unauthorized("User not found");
        }

        var result = await _obstaclesService.AddSolvedVoteAsync(user.Identifier, trailObstacleIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("AddSolvedVote: Failed to add solved vote for user: {userIdentifier}", user.Identifier);

            return ToActionResult(result.Message);
        }

        return Ok();
    }

    [Authorize]
    [HttpPut]
    [Route("{trailObstacleIdentifier}")]
    public async Task<ActionResult> UpdateTrailObstacle(
        [FromRoute] string trailObstacleIdentifier,
        TrailObstacleUpdateRequest updateRequest,
        CancellationToken ctoken)
    {
        var user = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (user is null)
        {
            return Unauthorized("User not found");
        }

        var result = await _obstaclesService.UpdateTrailObstacleAsync(
            user.Identifier,
            trailObstacleIdentifier,
            updateRequest.Description,
            updateRequest.IssueType,
            ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("UpdateTrailObstacle: Failed to update trail obstacle for user: {userIdentifier}", user.Identifier);
            return ToActionResult(result.Message);
        }

        return NoContent();
    }

    [Authorize]
    [HttpDelete]
    [Route("solve/{trailObstacleIdentifier}")]
    public async Task<ActionResult> DeleteSolvedVoteByUserIdentifier([FromRoute] string trailObstacleIdentifier, CancellationToken ctoken)
    {
        var user = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (user is null)
        {
            return Unauthorized("User not found");
        }

        var result = await _obstaclesService.DeleteSolvedVoteByUserIdentifierAsync(user.Identifier, trailObstacleIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation("DeleteSolvedVoteByUserIdentifier: Failed to delete solved vote for user: {userIdentifier}", user.Identifier);

            return ToActionResult(result.Message);
        }

        return NoContent();
    }

    [Authorize]
    [HttpDelete]
    [Route("{trailObstacleIdentifier}")]
    public async Task<ActionResult> DeleteTrailObstacle([FromRoute] string trailObstacleIdentifier, CancellationToken ctoken)
    {
        var user = await GetAuthenticatedUserAsync(_userService, ctoken);

        if (user is null)
        {
            return Unauthorized("User not found");
        }

        var result = await _obstaclesService.DeleteTrailObstacleAsync(user.Identifier, trailObstacleIdentifier, ctoken);

        if (!result.Success && result.Message != null)
        {
            _logger.LogInformation($"DeleteTrailObstacle: Failed to delete trail obstacle for user: {user.Identifier}");

            return ToActionResult(result.Message);
        }

        return NoContent();
    }
}