# web/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "StigVidd",
    "description": "Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/v1/Reviews/trail/{trailIdentifier}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "operationId": "Reviews_GetReviewsByTrailIdentifier",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedReviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/TrailObstacles/trail/{trailIdentifier}": {
      "get": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_GetTrailObstacles",
        "parameters": [
          {
            "name": "trailIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailObstacleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/TrailObstacles/issue-types": {
      "get": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_GetTrailObstacleIssueTypes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/TrailObstacles": {
      "post": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_CreateTrailObstacle",
        "requestBody": {
          "x-name": "obstacleRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailObstacleRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/TrailObstacles/solve/{trailObstacleIdentifier}": {
      "post": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_AddSolvedVote",
        "parameters": [
          {
            "name": "trailObstacleIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_DeleteSolvedVoteByUserIdentifier",
        "parameters": [
          {
            "name": "trailObstacleIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/TrailObstacles/{trailObstacleIdentifier}": {
      "put": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_UpdateTrailObstacle",
        "parameters": [
          {
            "name": "trailObstacleIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailObstacleUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "TrailObstacles"
        ],
        "operationId": "TrailObstacles_DeleteTrailObstacle",
        "parameters": [
          {
            "name": "trailObstacleIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Trails/{identifier}": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetTrailByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_UpdateTrail",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrailRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/Trails": {
      "get": {
        "tags": [
          "Trails"
        ],
        "operationId": "Trails_GetAllTrails",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailShortInfoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrailImageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "imageUrl"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedReviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "reviews"
        ],
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "hasMore": {
            "type": "boolean"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "trailIdentifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "trailReview": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "format": "decimal"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userIdentifier": {
            "type": "string",
            "nullable": true
          },
          "trailIdentifier": {
            "type": "string"
          },
          "reviewImages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ReviewImageResponse"
            }
          }
        }
      },
      "ReviewImageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "imageUrl"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          }
        }
      },
...
      "TrailObstacleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "description",
          "issueType"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "userIdentifier": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "issueType": {
            "type": "string"
          },
          "incidentLongitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "incidentLatitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "solvedVotes": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailObstacleSolvedVoteResponse"
            }
          }
        }
      },
      "TrailObstacleSolvedVoteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userIdentifier",
          "trailObstacleIdentifier"
        ],
        "properties": {
          "userIdentifier": {
            "type": "string"
          },
          "trailObstacleIdentifier": {
            "type": "string"
          }
        }
      },
      "TrailObstacleRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "trailIdentifier",
          "description",
          "issueType"
        ],
        "properties": {
          "trailIdentifier": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "issueType": {
            "type": "string"
          },
          "incidentLongitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "incidentLatitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
      "TrailObstacleUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "issueType": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TrailResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "trailLength": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "classification": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accessibility": {
            "type": "boolean",
            "nullable": true
          },
          "accessibilityInfo": {
            "type": "string",
            "nullable": true
          },
          "trailSymbol": {
            "type": "string",
            "nullable": true
          },
          "trailSymbolImage": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fullDescription": {
            "type": "string",
            "nullable": true
          },
          "coordinates": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "visitorInformation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/VisitorInformationResponse"
              }
            ]
          },
          "trailImagesResponse": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailImageResponse"
            }
          },
          "trailLinksResponse": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrailLinkResponse"
            }
          }
        }
      },
      "VisitorInformationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "gettingThere": {
            "type": "string"
          },
          "publicTransport": {
            "type": "string"
          },
          "parking": {
            "type": "string"
          },
          "illumination": {
            "type": "boolean"
          },
          "illuminationText": {
            "type": "string"
          },
          "maintainedBy": {
            "type": "string"
          },
          "winterMaintenance": {
            "type": "boolean"
          }
        }
      },
      "TrailLinkResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "link",
          "title"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "TrailShortInfoResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "identifier",
          "name",
          "trailLength",
          "city"
        ],
        "properties": {
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trailLength": {
            "type": "number",
            "format": "decimal"
          },
          "accessibility": {
            "type": "boolean"
          },
          "classification": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "startLatitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "startLongitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
...
      "UpdateTrailRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "trailLength"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "trailLength": {
            "type": "number",
            "format": "decimal"
          },
          "classification": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accessibility": {
            "type": "boolean",
            "nullable": true
          },
          "accessibilityInfo": {
            "type": "string",
            "nullable": true
          },
          "trailSymbol": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fullDescription": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "visitorInformation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/UpdateVisitorInformationRequest"
              }
            ]
          }
        }
      },
      "UpdateVisitorInformationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "gettingThere": {
            "type": "string",
            "nullable": true
          },
          "publicTransport": {
            "type": "string",
            "nullable": true
          },
          "parking": {
            "type": "string",
            "nullable": true
          },
          "illumination": {
            "type": "boolean",
            "nullable": true
          },
          "illuminationText": {
            "type": "string",
            "nullable": true
          },
          "maintainedBy": {
            "type": "string",
            "nullable": true
          },
          "winterMaintenance": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Skriv: Bearer {din Keycloak access token}",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}